Create Dice roll game
User can provide the max number of player between 2-4 for the Dice roll game. Max score is 50, if users score beyond or equal to max score the game will terminate. User score is cumulative total value of the Dice rolls, except when the Dice roll value is zero. Then the total score of the player will be reduced to zero.